fix: packing improvements

+++ src/PriceStorage.sol
@@ -7,40 +7,40 @@ import {IPriceStorage} from "./interfaces/IPriceStorage.sol";
 
 contract PriceStorage is IPriceStorage, AccessControlDefaultAdminRulesUpgradeable {
   bytes32 public constant SERVICE_ROLE = keccak256("SERVICE_ROLE");
-  uint256 public constant BOUND_PERCENTAGE_DENOMINATOR = 1e18;
+  uint128 public constant BOUND_PERCENTAGE_DENOMINATOR = 1e18;
 
   mapping(bytes32 key => Price price) public prices;
   Price public lastPrice;
 
-  uint256 public upperBoundPercentage;
-  uint256 public lowerBoundPercentage;
+  uint128 public upperBoundPercentage;
+  uint128 public lowerBoundPercentage;
 
   /// @custom:oz-upgrades-unsafe-allow constructor
   constructor() {
     _disableInitializers();
   }
 
-  function initialize(uint256 _upperBoundPercentage, uint256 _lowerBoundPercentage) public initializer {
+  function initialize(uint128 _upperBoundPercentage, uint128 _lowerBoundPercentage) public initializer {
     __AccessControlDefaultAdminRules_init(1 days, msg.sender);
     setUpperBoundPercentage(_upperBoundPercentage);
     setLowerBoundPercentage(_lowerBoundPercentage);
   }
 
-  function setPrice(bytes32 _key, uint256 _price) external onlyRole(SERVICE_ROLE) {
+  function setPrice(bytes32 _key, uint128 _price) external onlyRole(SERVICE_ROLE) {
     if (_key == bytes32(0)) revert InvalidKey();
     if (_price == 0) revert InvalidPrice();
     if (prices[_key].timestamp != 0) revert PriceAlreadySet(_key);
 
-    uint256 lastPriceValue = lastPrice.price;
+    uint128 lastPriceValue = lastPrice.price;
     if (lastPriceValue != 0) {
-      uint256 upperBound = lastPriceValue + ((lastPriceValue * upperBoundPercentage) / BOUND_PERCENTAGE_DENOMINATOR);
-      uint256 lowerBound = lastPriceValue - ((lastPriceValue * lowerBoundPercentage) / BOUND_PERCENTAGE_DENOMINATOR);
+      uint128 upperBound = lastPriceValue + ((lastPriceValue * upperBoundPercentage) / BOUND_PERCENTAGE_DENOMINATOR);
+      uint128 lowerBound = lastPriceValue - ((lastPriceValue * lowerBoundPercentage) / BOUND_PERCENTAGE_DENOMINATOR);
       if (_price > upperBound || _price < lowerBound) {
         revert InvalidPriceRange(_price, lowerBound, upperBound);
       }
     }
 
-    uint256 currentTime = block.timestamp;
+    uint128 currentTime = uint128(block.timestamp);
     Price memory price = Price({price: _price, timestamp: currentTime});
     prices[_key] = price;
     lastPrice = price;
@@ -48,7 +48,7 @@ contract PriceStorage is IPriceStorage, AccessControlDefaultAdminRulesUpgradeabl
     emit PriceSet(_key, _price, currentTime);
   }
 
-  function setUpperBoundPercentage(uint256 _upperBoundPercentage) public onlyRole(DEFAULT_ADMIN_ROLE) {
+  function setUpperBoundPercentage(uint128 _upperBoundPercentage) public onlyRole(DEFAULT_ADMIN_ROLE) {
     if (_upperBoundPercentage == 0 || _upperBoundPercentage > BOUND_PERCENTAGE_DENOMINATOR)
       revert InvalidUpperBoundPercentage();
 
@@ -56,7 +56,7 @@ contract PriceStorage is IPriceStorage, AccessControlDefaultAdminRulesUpgradeabl
     emit UpperBoundPercentageSet(_upperBoundPercentage);
   }
 
-  function setLowerBoundPercentage(uint256 _lowerBoundPercentage) public onlyRole(DEFAULT_ADMIN_ROLE) {
+  function setLowerBoundPercentage(uint128 _lowerBoundPercentage) public onlyRole(DEFAULT_ADMIN_ROLE) {
     if (_lowerBoundPercentage == 0 || _lowerBoundPercentage > BOUND_PERCENTAGE_DENOMINATOR)
       revert InvalidLowerBoundPercentage();
